Backend/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
numpy
scipy
PyYAML
pytest
httpx
python-multipart
sqlmodel
